db::hints: properly stop resource_manager
space_watchdog blocks on the end_point_manager::file_update_mutex().
Therefore when we shut scylla down we need to first stop the space_watchdog
(which is a part of resource_manager) and only after that shut all instances
of db::hints::manager down. Otherwise there may be a use-after-free event.

However we should also make sure that hints do not break the disk space
contract during the time frame when space_watchdog is down but
hints::managers are still up. In order to ensure that we will disable
hints storing before stopping space_watchdog.

Fixes #4836

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>
Message-Id: <20190812201857.5716-2-vladz@scylladb.com>