database: Avoid OOMing with flush continuations after failed memtable flush
The original fix (10f6b125c8c77b5914b4c7e52350639e027a989a) didn't
take into account that if there was a failed memtable flush (Refs
pick a candidate region which is not empty (failed the flush) but is
not a flushable memtable because it's not the latest in the memtable
list.

Suspected to be observed in #4982. I didn't manage to reproduce after
triggering a failed memtable flush.

Fixes #3717