database: Fix excessive flushing of small memtables in some degenerate cases
Before this patch, when memtable flushing fails we give back the
amount of virtual dirty which was consumed during flushing so that
when flushing is retried later it can start decreasing virtual dirty
again from scratch.

This means that the failed memtable will take up space in the dirty
region but is not flushable anymore (it will retry flushing after 10s
automatically). As a result, it will take much less writes to
accumulate in order for us to cross the soft pressure threshold
again. In particular, if only one table is active, we will be
constantly under soft pressure. As a result, if writes are still
incoming, we will keep flushing very small memtables.

This may quickly lead to OOM due to too many small sstables.

Refs #4235, #4116

The fix is to make sure that a memtable takes no space in dirty after
the permit is released. It is moved to real dirty entirely.

This also affects the non-failure flush case. Before this patch, there
was a time window between the permit was released and the cache update
was started in which the memtable was still taking space in the dirty
region, the more the sparser it was.

Fixes #4344

To illustrate how memtable flush decisions change, below is a comparison
of memtables present in memory during commitlog replay when it stops
due to real dirty:

Before:

(gdb) scylla memtables
table "keyspace1"."standard1":
  (memtable*) 0x6000029de480: total=145096704, used=31760496, free=113336208, evictable=0, flushed=30715790
  (memtable*) 0x60000017f680: total=100401152, used=71661352, free=28739800,  evictable=0, flushed=69555617
  (memtable*) 0x6000070e8480: total=91881472, used=70292352, free=21589120, evictable=0, flushed=68182174
  (memtable*) 0x60000256cf00: total=71958528, used=55353376, free=16605152, evictable=0, flushed=53684254
  (memtable*) 0x600000184300: total=53739520, used=41647080, free=12092440, evictable=0, flushed=40318984
  (memtable*) 0x60000256d080: total=20185088, used=15721784, free=4463304, evictable=0, flushed=15139755
  (memtable*) 0x6000070e8300: total=0, used=0, free=0, evictable=0, flushed=0
  ...

After:

(gdb) scylla memtables
table "keyspace1"."standard1":
  (memtable*) 0x600006282f00: total=145096704, used=31759936, free=113336768, evictable=0, flushed=0
  (memtable*) 0x600002b67800: total=144965632, used=103425840, free=41539792, evictable=0, flushed=0
  (memtable*) 0x600002b67b00: total=187170816, used=138431992, free=48738824, evictable=0, flushed=0
  (memtable*) 0x600002b67980: total=6029312, used=4759000, free=1270312, evictable=6029312, flushed=0

Note that in the After case, the last memtable is still active.

Tests:

   - unit (dev)
   - scylla-bench write workload to observe latency
   - commitlog replay with memtable flush failing