sstables: mx: bsearch_clustered_cursor: Split promoted index blocks cache from reader state
Currently, cached_promoted_index holds both the cache of parsed
entries and the state needed during population of the cache.

Separate the two in preparation for making the cache shared. The
shared part which holds the entries will is called
cached_promoted_index and the reader (not shared) which populates it
is called cached_promoted_index_reader.