migration_manager: Make schema pulls more robust
Currently schema pulls are edge-triggered in reaction to changes in
gossip status of the nodes. If schema pull fails for some reason, it
will not be retried until the next schema change is triggered.

To make schema synchronization more robust, this patch reworks the pull
mechanism to be level-triggered instead.

There will be a worker fiber which periodically checks advertised schema
versions and issues pulls if necessary.

The worker can also be woken-up from gossip status change listeners,
so that schema is synchronized quickly during bootstrapp, like before.

Fixes #7426.