tests: sstables: Fix test assertions to not expect more than they should
Before this patch, the tests expected readers to emit range tombstones
which are outside clustering restrictions. Readers do not have to emit
range tombstones outside clustering restrictions, so fix tests to only
expect the part which overlaps with query ranges.

This is a preparatory patch before changing readers to trim range
tombstones to clustering ranges.