row_cache: Fix undefined behavior during eviction under some conditions
When the last non-dummy row is evicted from a partition, the partition
entry is evicted as well. The existing logic in on_evicted() leaves
the last dummy row in the partition version before evicting the
partition entry. This row may still be attached to the LRU. Eviction
of partition entry goes through mutation_cleaner::clear_gently(). If
this is preempted, the destruction may proceed in the background. If
evicition happens on the remaining row in that entry before it's
destroyed, the code will hit undefined behavior. on_evicted() calls
partition_version::is_referenced_from_entry(), which is unspecified
when the version is enqueued in the mutation_cleaner. It returns
incorrect value for the last item remaining in the LRU. In that case
eviction will try to access non-existent containing partition_entry,
causing undefined behavior.

Caught by debug-mode cql_query_test.test_clustering_filtering with
raft enabled. Where it manifested like this:

  partition_version.hh:328:16: runtime error: load of value 7, which is not a valid value for type 'bool'
  SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior partition_version.hh:328:16 in
  Aborting on shard 0.

Instances of this issue outside of the unit test environment are not
known as of yet.

This change makes is_referenced_from_entry() return the correct value
even for versions which are queued in the mutation cleaner.

Fixes #11140