row_cache: Do not leave garbage rows in the partition entry prior to evicting it
Just an optimization / cleanup which makes sure row entries which are
still linked in the LRU don't land in the versions queued in the
mutation cleaner. Eviction from such versions is oficially prohibited,
although in the case of versions with only a single dummy row it does
no harm. Still, let's respect the formal requirement. It may also save
some work in case preemption is signalled and we're evicting lots of
partitions.