test: tablets_test: Add support for auto-split mode
rebalance_tablets() was performing migraitons and merges automatically
but not splits, because splits need to be acked by replicas via
load_stats. It's inconvienient in tests which want to rebalance to the
equilibrium point. This patch changes rebalance_tablets() to split
automatically by default, can be disabled for tests which expect
differently.

shared_load_stats was introduced to provide a stable holder of
load_stats which can be reused across rebalance_tablets() calls.